Add files via upload
This is one of the first 3D prototypes based on my MonoPhan Project experience for the past 10 years. fully 3D printed and assembled with epoxy.  This is a 3 x 50mm EDF setup with a rear thrust vectoring nozzle for yaw.
I printed all parts with standard PLA with standard speed and layer settings in Cura. You need to rotate them around and print them at the correct position for the support, etc. based on your preferences. Thanks